+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
-## [Unreleased]
+## [9.6.0] - 2025-07-24
 ### Added
 * Add `REAL_TIME_UPDATES_TOPICS` config option.
 * Add CORS-related config options.